+++ tests/core/ibc/client/grpc/test_ibc_client_grpc_api.py
@@ -5,7 +5,7 @@ import pytest
 from google.protobuf import any_pb2
 
 from pyinjective.client.model.pagination import PaginationOption
-from pyinjective.core.ibc.client.ibc_client_grpc_api import IBCClientGrpcApi
+from pyinjective.core.ibc.client.grpc.ibc_client_grpc_api import IBCClientGrpcApi
 from pyinjective.core.network import DisabledCookieAssistant, Network
 from pyinjective.proto.cosmos.base.query.v1beta1 import pagination_pb2 as pagination_pb
 from pyinjective.proto.cosmos.ics23.v1 import proofs_pb2 as ics23_proofs